Ajustando chamada

+++ app/src/main/java/com/br/repogit/domain/model/RepositoryDomain.kt
@@ -1,12 +1,12 @@
 package com.br.repogit.domain.model
 
 data class RepositoryDomain(
-    val id: Long,
+    val id: Int,
     val name: String,
     val fullName: String,
     val private: Boolean,
     val owner: OwnerDomain,
     val description: String? = null,
-    val forksCount: Long,
-    val stargazersCount: Long
+    val forksCount: Int,
+    val stargazersCount: Int
 )
