+++ app/src/main/java/com/br/repogit/presentation/model/RepositoryPresentation.kt
@@ -1,12 +1,12 @@
 package com.br.repogit.presentation.model
 
 data class RepositoryPresentation(
-    val id: Long,
+    val id: Int,
     val name: String,
     val fullName: String,
     val private: Boolean,
     val owner: OwnerPresentation,
     val description: String? = null,
-    val forksCount: Long,
-    val stargazersCount: Long
+    val forksCount: Int,
+    val stargazersCount: Int
 )
